Revert last change, kind of, re: getSnapshot always showing latest
It made the code more complex, and in practice, the whole reason we
called it `getSnapshot` is so that people know it might be stale.

This sort of code should be pretty rare:

liveDst.observe { liveSrc.get() }
liveSrc.set("new value")
assertThat(liveDst.getSnapshot()).equals("new value") // FAIL

and so it's better to keep the code simple and document that the above
case is intended.

If users need a way to do that, we'll find a way to add it later, maybe
even something like

liveDst.refreshSnapshot()

but we'll see....